+++ pgp-engine/src/main/java/ltechnologies/onionphone/pgpshield/engine/UserIdManager.kt
@@ -73,7 +73,7 @@ class UserIdManager {
                 setSignatureCreationTime(true, now)
             }.generate()
             val sigGen = PGPSignatureGenerator(
-                PgpOperators.contentSignerBuilder(publicKey.algorithm, useBc),
+                PgpOperators.contentSignerBuilder(publicKey, useBc),
             )
             sigGen.setHashedSubpackets(hashed)
             sigGen.init(PGPSignature.CERTIFICATION_REVOCATION, privateKey)
@@ -134,7 +134,7 @@ class UserIdManager {
             setPrimaryUserID(false, primary)
         }.generate()
         val sigGen = PGPSignatureGenerator(
-            PgpOperators.contentSignerBuilder(publicKey.algorithm, useBc),
+            PgpOperators.contentSignerBuilder(publicKey, useBc),
         )
         sigGen.setHashedSubpackets(hashed)
         sigGen.init(PGPSignature.POSITIVE_CERTIFICATION, privateKey)